HospitalService/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.2.0
uvicorn==0.30.1
djangorestframework-camel-case==1.4.2
drf-yasg==1.21.8
whitenoise==6.7.0


requests
